refactor: 電力接続の範囲計算サービスをUtil/ElectricWireへ移設
