--- moorestech_server/Assets/Scripts/Game.World.EventHandler/EnergyEvent.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: eed679bbf44b8e14aa3b5264ce1bdc2c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
