--- moorestech_server/Assets/Scripts/Game.World.EventHandler/EnergyEvent/EnergyService.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f29ebcdc331794c4fb4b922b847c9378
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
